my_Chainlit/pyproject.toml
[project]
name = "my-chainlit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Yasir Zoha", email = "zyasir491@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "chainlit>=2.4.1",
    "openai-agents>=0.0.4",
    "python-dotenv>=1.0.1",
]

[project.scripts]
my-chainlit = "my_chainlit:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
